Disable UpdateExpression from Chapter 3 (#76)
* Disable UpdateExpression from Chapter 3

This is not a specified operator in the documentation

* move disalloed syntax to new file

* fix tsc errors

* Bump version
