+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "js-slang",
-  "version": "0.1.8",
+  "version": "0.1.9",
   "description": "Javascript-based interpreter for slang, written in Typescript",
   "author": {
     "name": "Source Academy",
