+++ src/__tests__/block_scoping.ts
@@ -154,43 +154,3 @@ test("Cannot overwrite loop variables within a block", () => {
   });
 });
 
-test("Cannot leave blank expressions in for loops", () => {
-  const code = [
-    `
-    for(; i < 3; i++) {
-      break;
-    }
-  `,
-    `
-    for(let i = 0; ; i++) {
-      break;
-    }
-  `,
-    `
-    for(let i = 0; i < 3;) {
-      break;
-    }
-  `,
-    `
-    for(;;) {
-      break;
-    }
-  `
-  ];
-  const scheduler = "preemptive";
-  const promises = code.map(c => {
-    const context = mockContext(3);
-    return runInContext(c, context, { scheduler }).then(obj => ({
-      context,
-      obj
-    }));
-  });
-  return Promise.all(promises).then(results => {
-    results.map(res => {
-      const { context, obj } = res;
-      expect(obj.status).toBe("error");
-      const errors = parseError(context.errors);
-      expect(errors).toMatchSnapshot();
-    });
-  });
-});
